Deye_cloud/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pymongo>=4.6.1
pydantic>=2.6.1